Split the base registration page with local password registration
