+++ templates/pages/login.html
@@ -11,7 +11,7 @@ Please see LICENSE in the repository root for full details.
 {% from "components/idp_brand.html" import logo %}
 
 {% block content %}
-  <main class="flex flex-col gap-10">
+  <form method="POST" class="flex flex-col gap-10">
     <header class="page-heading">
       <div class="icon">
         {{ icon.user_profile_solid() }}
@@ -31,54 +31,52 @@ Please see LICENSE in the repository root for full details.
       {% endif %}
     </header>
 
-    <form method="POST" class="flex flex-col gap-10">
-      <div class="cpd-form-root">
-        {% if form.errors is not empty %}
-          {% for error in form.errors %}
-            <div class="text-critical font-medium">
-              {{ errors.form_error_message(error=error) }}
-            </div>
-          {% endfor %}
-        {% endif %}
+    <div class="cpd-form-root">
+      {% if form.errors is not empty %}
+        {% for error in form.errors %}
+          <div class="text-critical font-medium">
+            {{ errors.form_error_message(error=error) }}
+          </div>
+        {% endfor %}
+      {% endif %}
 
-        <input type="hidden" name="csrf" value="{{ csrf_token }}" />
+      <input type="hidden" name="csrf" value="{{ csrf_token }}" />
 
-        {% call(f) field.field(label=_("common.username"), name="username", form_state=form) %}
-          <input {{ field.attributes(f) }} class="cpd-text-control" type="text" autocomplete="username" autocorrect="off" autocapitalize="off" required />
+      {% call(f) field.field(label=_("common.username"), name="username", form_state=form) %}
+        <input {{ field.attributes(f) }} class="cpd-text-control" type="text" autocomplete="username" autocorrect="off" autocapitalize="off" required />
+      {% endcall %}
+
+      {% if features.password_login %}
+        {% call(f) field.field(label=_("common.password"), name="password", form_state=form) %}
+          <input {{ field.attributes(f) }} class="cpd-text-control" type="password" autocomplete="password" required />
         {% endcall %}
 
-        {% if features.password_login %}
-          {% call(f) field.field(label=_("common.password"), name="password", form_state=form) %}
-            <input {{ field.attributes(f) }} class="cpd-text-control" type="password" autocomplete="password" required />
-          {% endcall %}
-
-          {% if features.account_recovery %}
-            {{ button.link_text(text=_("mas.login.forgot_password"), href="/recover", class="self-center") }}
-          {% endif %}
+        {% if features.account_recovery %}
+          {{ button.link_text(text=_("mas.login.forgot_password"), href="/recover", class="self-center") }}
         {% endif %}
-      </div>
+      {% endif %}
+    </div>
 
-      <div class="cpd-form-root">
-        {% if features.password_login %}
-          {{ button.button(text=_("action.continue")) }}
-        {% endif %}
+    <div class="cpd-form-root">
+      {% if features.password_login %}
+        {{ button.button(text=_("action.continue")) }}
+      {% endif %}
 
-        {% if features.password_login and providers %}
-          {{ field.separator() }}
-        {% endif %}
+      {% if features.password_login and providers %}
+        {{ field.separator() }}
+      {% endif %}
 
-        {% if providers %}
-          {% set params = next["params"] | default({}) | to_params(prefix="?") %}
-          {% for provider in providers %}
-            {% set name = provider.human_name or (provider.issuer | simplify_url(keep_path=True)) or provider.id %}
-            <a class="cpd-button {%- if provider.brand_name %} has-icon {%- endif %}" data-kind="secondary" data-size="lg" href="{{ ('/upstream/authorize/' ~ provider.id ~ params) | prefix_url }}">
-              {{ logo(provider.brand_name) }}
-              {{ _("mas.login.continue_with_provider", provider=name) }}
-            </a>
-          {% endfor %}
-        {% endif %}
-      </div>
-    </form>
+      {% if providers %}
+        {% set params = next["params"] | default({}) | to_params(prefix="?") %}
+        {% for provider in providers %}
+          {% set name = provider.human_name or (provider.issuer | simplify_url(keep_path=True)) or provider.id %}
+          <a class="cpd-button {%- if provider.brand_name %} has-icon {%- endif %}" data-kind="secondary" data-size="lg" href="{{ ('/upstream/authorize/' ~ provider.id ~ params) | prefix_url }}">
+            {{ logo(provider.brand_name) }}
+            {{ _("mas.login.continue_with_provider", provider=name) }}
+          </a>
+        {% endfor %}
+      {% endif %}
+    </div>
 
     {% if (not next or next.kind != "link_upstream") and features.password_registration %}
       <div class="flex gap-1 justify-center items-center cpd-text-body-md-regular">
@@ -91,21 +89,10 @@ Please see LICENSE in the repository root for full details.
       </div>
     {% endif %}
 
-    {% if next and next.kind == "continue_authorization_grant" %}
-      {{ back_to_client.link(
-        text=_("action.cancel"),
-        kind="secondary",
-        destructive=True,
-        uri=next.grant.redirect_uri,
-        mode=next.grant.response_mode,
-        params=dict(error="access_denied", state=next.grant.state)
-      ) }}
-    {% endif %}
-
     {% if not providers and not features.password_login %}
       <div class="text-center">
         {{ _("mas.login.no_login_methods") }}
       </div>
     {% endif %}
-  </main>
+  </form>
 {% endblock content %}
